+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## Unreleased
 
+- Add support for `inspect sandbox cleanup k8s` command to uninstall all Inspect Helm charts.
 - Initial release.